refactor(ssg): :art: renaming term to cmd
